// src/models/menu.ts
import { SvgIconComponent } from '@mui/icons-material';

export interface IManagementMenu {
  id: number;
  content: string;
  link: string;
  icon: JSX.Element;
}

export interface ICreateMenu {
  id: number;
  content: string;
  link: string;
  icon: JSX.Element;
}

export interface IEditMenu {
  id: number;
  content: string;
  link: string;
  icon: JSX.Element;
}

export interface IDetailMenu {
  id: number;
  content: string;
  link: string;
  icon: JSX.Element;
}